refactor declarations
