+++ include/boost/http_proto/src.hpp
@@ -32,6 +32,7 @@ in a translation unit of the program.
 #include <boost/http_proto/impl/field.ipp>
 #include <boost/http_proto/impl/fields.ipp>
 #include <boost/http_proto/impl/fields_view.ipp>
+#include <boost/http_proto/impl/fields_view_base.ipp>
 #include <boost/http_proto/impl/header_info.ipp>
 #include <boost/http_proto/impl/method.ipp>
 #include <boost/http_proto/impl/parser.ipp>
